+++ rewatch/testrepo/packages/dep02/src/Array.mjs
@@ -291,9 +291,13 @@ let get = Belt_Array.get;
 
 let getExn = Belt_Array.getExn;
 
+let getOrThrow = Belt_Array.getOrThrow;
+
 let set = Belt_Array.set;
 
 let setExn = Belt_Array.setExn;
+
+let setOrThrow = Belt_Array.setOrThrow;
 
 let shuffleInPlace = Belt_Array.shuffleInPlace;
 
@@ -434,8 +438,10 @@ let sortByRaw = Belt_SortArray.stableSortBy;
 export {
   get,
   getExn,
+  getOrThrow,
   set,
   setExn,
+  setOrThrow,
   shuffleInPlace,
   shuffle,
   reverseInPlace,